requirements.txt
asgiref==3.3.1
celery==5.0.5
Django==2.2.16
django-cors-headers==3.7.0
django_celery_beat==2.2.0
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-rest-framework-mongoengine==3.4.1
django-simple-captcha==0.5.13
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-extensions==0.7.0
jsonpath==0.82
mongoengine==0.22.1
mysqlclient==2.0.3
Pillow==10.0.0
PyJWT==1.7.1
pymongo==3.11.3
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
xlwt==1.3.0
xlrd==2.0.1
coreapi==2.3.3
user-agents==2.2.0
eventlet==0.30.2
psutil==5.8.0
drf-yasg==1.20.0
djangorestframework-simplejwt=5.3.1
django_restql=0.15.3
qrcode=7.4.2
PyYAML=6.0.1
attrs=23.2.0
drf-spectacular=0.27.0
jsonschema=4.20.0
jsonschema-specifications=2023.12.1
referencing=0.32.0
rpds-py=0.16.2
drf-spectacular-sidecar=2024.1.1